fix: adjusted packages
