+++ clients/spring-boot-starter-camunda-sdk/src/main/java/io/camunda/zeebe/spring/client/properties/common/ZeebeClientProperties.java
@@ -18,6 +18,7 @@ package io.camunda.zeebe.spring.client.properties.common;
 import io.camunda.zeebe.spring.client.annotation.value.ZeebeWorkerValue;
 import java.net.URI;
 import java.time.Duration;
+import java.util.HashMap;
 import java.util.Map;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
 
@@ -30,8 +31,8 @@ public class ZeebeClientProperties extends ApiProperties {
   private String caCertificatePath;
   private Duration keepAlive;
   private String overrideAuthority;
-  @NestedConfigurationProperty private ZeebeWorkerValue defaults;
-  @NestedConfigurationProperty private Map<String, ZeebeWorkerValue> override;
+  @NestedConfigurationProperty private ZeebeWorkerValue defaults = new ZeebeWorkerValue();
+  @NestedConfigurationProperty private Map<String, ZeebeWorkerValue> override = new HashMap<>();
   private boolean preferRestOverGrpc;
   private URI grpcAddress;
   private URI restAddress;
